fastapi-simple-architecture/pyproject.toml
[project]
name = "simple-layer-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pact-python>=3.2.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
